requirements.txt
fastapi
uvicorn[standard]
python-dotenv
oracledb
pydantic
PyJWT
bcrypt
python-multipart
typing-extensions
